Made some general stricture changes
Moved all apps to an single app for current game
