--- backend/yugiohcardinset/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class YugiohcardinsetConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'yugiohcardinset'
-    verbose_name = 'Yugioh Card In Set'
