--- backend/yugiohcardrarity/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class YugiohcardrarityConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'yugiohcardrarity'
-    verbose_name = 'Yugioh Card Rarity'
