--- backend/yugiohcardset/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class YugiohcardsetConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'yugiohcardset'
-    verbose_name = "Yugioh Card Set"
